requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
greenlet==3.3.2
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.24.1
jsonschema-path==0.4.2
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openapi-schema-validator==0.7.2
openapi-spec-validator==0.8.1
packaging==26.0
pathable==0.5.0
playwright==1.58.0
prance==25.4.8.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyee==13.0.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.3.3
-e git+https://github.com/ayushkumar2601/Rico-private.git@bd14b6f359c4544eff07a50cc3af13cbd0eb7e55#egg=rico_sec
rpds-py==0.30.0
ruamel.yaml==0.19.1
shellingham==1.5.4
six==1.17.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
